调整代码，支持dubbo-2.6.0，调整artifactId
