+++ .gitignore
@@ -1,4 +1,6 @@
-/target
-/.settings
-/.project
-/.classpath
+target/
+.settings/
+.project
+.classpath
+*iml
+.idea
