--- src/main/java/com/alibaba/boot/dubbo/annotation/EnableDubboConfiguration.java
@@ -1,21 +1,0 @@
-package com.alibaba.boot.dubbo.annotation;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Enable Dubbo (for provider or consumer) for spring boot application
- *
- * @author xionghui
- * @email xionghui.xh@alibaba-inc.com
- * @since 1.0.0
- */
-@Target({ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface EnableDubboConfiguration {
-
-}
