--- src/main/java/com/alibaba/boot/dubbo/domain/SpringBootStarterDobboConstants.java
@@ -1,16 +1,0 @@
-package com.alibaba.boot.dubbo.domain;
-
-/**
- * 常量定义
- *
- * @author xionghui
- * @email xionghui.xh@alibaba-inc.com
- * @since 1.0.0
- */
-public class SpringBootStarterDobboConstants {
-
-  public static final String GROUP = "group";
-
-  public static final String VERSION = "version";
-
-}
