+++ src/main/java/com/alibaba/dubbo/spring/boot/annotation/EnableDubboConfiguration.java
@@ -1,0 +1,21 @@
+package com.alibaba.dubbo.spring.boot.annotation;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Enable Dubbo (for provider or consumer) for spring boot application
+ * 
+ * @author xionghui
+ * @version 1.0.0
+ * @since 1.0.0
+ */
+@Target({ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface EnableDubboConfiguration {
+
+}
