+++ src/main/java/com/alibaba/dubbo/spring/boot/domain/DubboSpringBootStarterConstants.java
@@ -1,0 +1,15 @@
+package com.alibaba.dubbo.spring.boot.domain;
+
+/**
+ * 常量定义
+ *
+ * @author xionghui
+ * @version 1.0.0
+ * @since 1.0.0
+ */
+public class DubboSpringBootStarterConstants {
+
+  public static final String GROUP = "group";
+
+  public static final String VERSION = "version";
+}
