+++ src/main/resources/META-INF/dubbo/com.alibaba.dubbo.rpc.ExporterListener
@@ -1,1 +1,1 @@
-dubboexport=com.alibaba.boot.dubbo.listener.ProviderExportListener
+dubboexport=com.alibaba.dubbo.spring.boot.listener.ProviderExportListener
