+++ src/main/resources/META-INF/dubbo/com.alibaba.dubbo.rpc.InvokerListener
@@ -1,1 +1,1 @@
-dubbosubscribe=com.alibaba.boot.dubbo.listener.ConsumerSubscribeListener
+dubbosubscribe=com.alibaba.dubbo.spring.boot.listener.ConsumerSubscribeListener
