+++ src/main/resources/META-INF/spring.factories
@@ -1,4 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.alibaba.boot.dubbo.DubboAutoConfiguration,\
-com.alibaba.boot.dubbo.DubboProviderAutoConfiguration,\
-com.alibaba.boot.dubbo.DubboConsumerAutoConfiguration
+com.alibaba.dubbo.spring.boot.DubboAutoConfiguration,\
+com.alibaba.dubbo.spring.boot.DubboProviderAutoConfiguration,\
+com.alibaba.dubbo.spring.boot.DubboConsumerAutoConfiguration
