+++ src/main/resources/META-INF/spring.provides
@@ -1,1 +1,1 @@
-provides: spring-boot-starter-dubbo
+provides: dubbo-spring-boot-starter
