--- src/test/java/com/alibaba/boot/dubbo/service/IHelloService.java
@@ -1,5 +1,0 @@
-package com.alibaba.boot.dubbo.service;
-
-public interface IHelloService {
-  public String hello();
-}
