+++ src/test/java/com/alibaba/dubbo/spring/boot/Mock.java
@@ -1,0 +1,5 @@
+package com.alibaba.dubbo.spring.boot;
+
+public class Mock {
+
+}
