--- src/test/resources/application.properties
@@ -1,16 +1,0 @@
-# 服务端口
-server.port=7001
-
-# tomcat最大线程数，默认为200
-server.tomcat.max-threads=10
-
-spring.http.encoding.charset=UTF-8
-spring.http.encoding.enabled=true
-
-# dubbo配置
-spring.dubbo.appname=spring-boot-starter-dubbo-test
-spring.dubbo.protocol=dubbo
-spring.dubbo.registry=multicast://224.0.0.0:1111
-spring.dubbo.port=20801
-spring.dubbo.group=test
-spring.dubbo.version=1.0.0
